requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
djangorestframework==3.11.1
djangorestframework-gis==0.15
Markdown==3.2.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
